move loss/optimizer helper methods
